Fixed linker errors by deleting all include paths and using auto inherit
I deleted everything in the include and lib paths for debug
configuration. Then I just clicked "inherit from parent or project
defaults" and now it compiles. No idea why the hell that fixed it.
Because it looks to me like the project defaults dont even know about
the old directx 2010 folder. One other thing I added was d3dcompiler.lib
to the linker path - for both debug and release. so maybe its only using
the deprecated compiler for the shaders, but nothing else?