Got DDS working - through HELL AND BACK
After much hell, got multitexturing with DDS working. Next step is to
organize the different multitex functions to use loops and manage
different numbers of textures in a clean way.
CAUTION:
1. Currently using a RAW ARRAY of ID3D11ShaderResourceView* pointers
called _textureViewsDDS[6];
I didnt do any cleanup for this; I had to use it because the
ShaderSettings function further down the line wants a
ID3D11ShaderResourceView** and it simply wont take an array of Com
pointers as a substitute. So I dont know, maybe have to manually loop
through the comptr array and create another array to pass to that
function?