Implemented Generic Render to water shader
2 new issues have come to light:
1. Currently basing texture (resource) array size on number of file
names being passed in to initialize material, but may have to RESERVE
EXTRA VIEW SLOTS in the array if we are using render textures. Currently
hacked this by just initializing it w 2 extra file names, but this is a
bad way of course, it wastefully initializes two more textures that wont
be used.
2. Some texture views need to be updated every frame (render textures
for reflection and refraction). Currently just re-pointing the pointers
inside the textureArray held by the model material each frame before
rendering it w the shader, but this is obviously bad.

TODO:
1. Implement this for all the other shaders so they all use the generic
Render call.
2. See what other materials need to be dynamically updated like teh
refraction adn reflection one.
3. Think of architecture to handle the above 2 issues (preallocating
empty slots in teh material for render targets, and how to update it
each frame)