[UNIT TESTING] Shadow bug is SOMEWHERE inside Graphics class
Still no idea why, painstakingly went through all the code in
graphics.cpp and compared it to the working version today, seems
identical. However, all of the other classes have been restored to old
working versions, so they cant be the cause of the problem.
***Only way forward***
is to start w the old stable code in graphics.cpp and reintroduce the
updated features into it manually one by one until the bug emerges
again.
TODO:
1. compare ALL includes of stable Graphics.h - maybe I missed some
possible causes of the bug.
2. After checking these, proceed with above plan.